fix convars and ActionCommands, add AttackLoop & UseLoop commands

+++ src/main/java/dev/aknologia/konsole/convar/GammaConVar.java
@@ -3,14 +3,14 @@ package dev.aknologia.konsole.convar;
 import dev.aknologia.konsole.niflheim.ConsoleVariable;
 import net.minecraft.client.MinecraftClient;
 
-public class GammaConVar extends ConsoleVariable<Integer> {
+public class GammaConVar extends ConsoleVariable<Double> {
     public GammaConVar() {
-        super("gamma", "Set the brightness setting.", 0, Integer.class);
+        super("gamma", "Set the brightness setting.", 0.0, Double.class);
     }
 
     @Override
-    public Integer get() {
-        return (int) MinecraftClient.getInstance().options.gamma;
+    public Double get() {
+        return MinecraftClient.getInstance().options.gamma;
     }
 
     @Override
